requirements.txt
streamlit==1.37.0
tensorflow==2.16.1
pillow==10.4.0
numpy==1.26.4